+++ src/main/java/com/buuz135/functionalstorage/inventory/CompactingInventoryHandler.java
@@ -147,7 +147,13 @@ public abstract class CompactingInventoryHandler implements IItemHandler, INBTSe
         if (isCreative()) return Integer.MAX_VALUE;
         if (slot == this.slots) return Integer.MAX_VALUE;
         int total = totalAmount;
-        if (hasDowngrade()) total = 64 * 9 * 9;
+        if (hasDowngrade()) {
+            if (this.slots == 2){
+                total = 64 * 9;
+            } else {
+                total = 64 * 9 * 9;
+            }
+        }
         return (int) Math.min(Integer.MAX_VALUE, Math.floor(
 		        (double) (total * (long) getMultiplier()) / this.resultList.get(slot).getNeeded()));
     }
@@ -155,7 +161,13 @@ public abstract class CompactingInventoryHandler implements IItemHandler, INBTSe
     public int getSlotLimitBase(int slot) {
         if (slot == this.slots) return Integer.MAX_VALUE;
         int total = totalAmount;
-        if (hasDowngrade()) total = 64 * 9 * 9;
+        if (hasDowngrade()) {
+            if (this.slots == 2){
+                total = 64 * 9;
+            } else {
+                total = 64 * 9 * 9;
+            }
+        }
         return (int) Math.min(Integer.MAX_VALUE, Math.floor(total / this.resultList.get(slot).getNeeded()));
     }